services custom normalizer and nelmio doc
