+++ composer.json
@@ -41,7 +41,9 @@
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^3.60",
-        "symfony/maker-bundle": "^1.60"
+        "symfony/maker-bundle": "^1.60",
+        "symfony/stopwatch": "7.1.*",
+        "symfony/web-profiler-bundle": "7.1.*"
     },
     "config": {
         "allow-plugins": {
