+++ config/packages/framework.yaml
@@ -1,9 +1,10 @@
 # see https://symfony.com/doc/current/reference/configuration/framework.html
 framework:
-    secret: '%env(APP_SECRET)%'
+    secret: "%env(APP_SECRET)%"
     #csrf_protection: true
-
-    # Note that the session will be started ONLY if you read or write from it.
+    profiler:
+        collect: true
+        # Note that the session will be started ONLY if you read or write from it.
     session: true
 
     #esi: true
