+++ config/routes.yaml
@@ -4,6 +4,4 @@ controllers:
         namespace: App\Controller
     type: attribute
 
-    # config/routes.yaml
-api_login_check:
-    path: /api/login_check
+
