+++ src/Controller/ProductController.php
@@ -19,31 +19,31 @@ use Symfony\Contracts\Cache\TagAwareCacheInterface;
 class ProductController extends AbstractController
 {
     /**
- * Cette méthode permet de récupérer l'ensemble des smartphones Bilemo.
- *
- * @OA\Response(
- *     response=200,
- *     description="Retourne la liste complète des smartphones Bilemo",
- *     @OA\JsonContent(
- *        type="array",
- *        @OA\Items(ref=@Model(type=Product::class, groups={"getCollection"}))
- *     )
- * )
- * @OA\Parameter(
- *     name="page",
- *     in="query",
- *     description="La page que l'on veut récupérer",
- *     @OA\Schema(type="integer")
- * )
- * @OA\Parameter(
- *     name="limit",
- *     in="query",
- *     description="Le nombre d'éléments que l'on veut récupérer",
- *     @OA\Schema(type="integer")
- * )
- * @OA\Tag(name="Products")
- * @Security(name="Bearer")
- */
+     * Cette méthode permet de récupérer l'ensemble des smartphones Bilemo.
+     *
+     * @OA\Response(
+     *     response=200,
+     *     description="Retourne la liste complète des smartphones Bilemo",
+     *     @OA\JsonContent(
+     *        type="array",
+     *     @Model(type=Product::class, groups={"getCollection"}))
+     *     )
+     * )
+     * @OA\Parameter(
+     *     name="page",
+     *     in="query",
+     *     @OA\Property(description="La page que l'on veut récupérer")
+     *     @OA\Schema(type="integer")
+     * )
+     * @OA\Parameter(
+     *     name="limit",
+     *     in="query",
+     *     description="Le nombre d'éléments que l'on veut récupérer",
+     *     @OA\Schema(type="integer")
+     * )
+     * @OA\Tag(name="Products")
+     * @Security(name="Bearer")
+     */
     #[Route('/api/products', name: 'get_collection', methods: ['GET'])]
     public function getCollection(
         ProductRepository $productRepository,
@@ -73,7 +73,17 @@ class ProductController extends AbstractController
             true
         );
     }
-
+    /**
+     * Récupérer un produit spécifique par son ID.
+     *
+     * @OA\Response(
+     *     response=200,
+     *     description="Retourne les détails d'un produit spécifique",
+     *     @OA\JsonContent(ref=@Model(type=Product::class, groups={"getItem"}))
+     * )
+     * @OA\Tag(name="Products")
+     * @Security(name="Bearer")
+     */
     #[Route('/api/products/{id}', name: 'get_item', methods: ['GET'])]
     public function getItem(
         int $id,
@@ -94,8 +104,9 @@ class ProductController extends AbstractController
             throw new NotFoundHttpException('Product not found');
         }
 
-        $jsonProduct = $serializer->serialize($product, 'json', ['groups' => 'getItem','item_operation_name' => true ]);
+        $jsonProduct = $serializer->serialize($product, 'json', ['groups' => 'getItem', 'item_operation_name' => true]);
 
         return new JsonResponse($jsonProduct, Response::HTTP_OK, [], true);
     }
+    
 }
