+++ src/Entity/Product.php
@@ -17,50 +17,62 @@ class Product
     #[Groups(['getCollection', 'getItem'])]
     private ?int $id = null;
 
+    
     #[ORM\Column(length: 255)]
     #[Groups(['getCollection', 'getItem'])]
     private ?string $name = null;
 
+    
     #[ORM\Column(length: 255)]
     #[Groups(['getCollection', 'getItem'])]
     private ?string $description = null;
 
+   
     #[ORM\Column(length: 255)]
     #[Groups(['getCollection', 'getItem'])]
     private ?string $brand = null;
 
+    
     #[ORM\Column(type: 'decimal', precision: 10, scale: 2)]
     #[Groups(['getCollection', 'getItem'])]
     private ?string $model = null;
 
+    
     #[ORM\Column]
     #[Groups(['getCollection', 'getItem'])]
     private ?int $price = null;
 
+    
     #[ORM\Column]
     #[Groups(['getCollection', 'getItem'])]
     private ?int $stock = null;
 
+   
     #[ORM\Column(type: 'datetime')]
     #[Groups(['getCollection', 'getItem'])]
     private ?\DateTimeInterface $dateAdded = null;
 
+   
     #[ORM\Column(nullable: true)]
     #[Groups(['getCollection', 'getItem'])]
     private ?array $technicalSpecs = null;
 
+    
     #[ORM\Column(nullable: true)]
     #[Groups(['getCollection', 'getItem'])]
     private ?array $images = null;
 
+   
     #[ORM\Column(length: 255, nullable: true)]
     #[Groups(['getCollection', 'getItem'])]
     private ?string $category = null;
 
+   
     #[ORM\Column(nullable: true)]
     #[Groups(['getCollection', 'getItem'])]
     private ?array $availableColors = null;
 
+   
     #[ORM\Column(length: 50, nullable: true)]
     #[Groups(['getCollection', 'getItem'])]
     private ?string $state = null;
