+++ src/Repository/ProductRepository.php
@@ -51,4 +51,5 @@ class ProductRepository extends ServiceEntityRepository
     //            ->getOneOrNullResult()
     //        ;
     //    }
+    
 }
