+++ templates/home/index.html.twig
@@ -1,0 +1,13 @@
+{% extends 'base.html.twig' %}
+{% block body %}
+    <section>
+        <h1>BileMo</h1>
+        <p><strong>Votre Api Bilemo</strong></p>
+        <p> 
+            <a href="/api/doc">Voir la documentation</a>
+        </p>
+        <footer>
+            <p><em>BileMo - Ludivine Goldstein - 2024</em></p>
+        </footer>
+    </section>
+{% endblock %}
